pyproject.toml
[tool.poetry]
name = "engenharia-de-dados"
version = "0.1.0"
description = "\"Estudando o módulo Engenharia de Dados na ALpha EdTech.\""
authors = ["Janaína Scher <janainascher@protonmail.com>"]
readme = "README.md"
packages = [{include = "engenharia_de_dados"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.0.1"
google-cloud-storage = "^2.9.0"
fsspec = "^2023.5.0"
gcsfs = "^2023.5.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
